feat: pass module dir as --path-prefix arg to golangci-lint
This ensures that the path output reported by golangci-lint is qualified
with respect to a given linted file's path within the repo. Without this
option, just the filename itself is outputted, leading to ambiguity as
to which file is problematic when multiple files have the same name
(e.g. `main.go`).